feat(function-container): store functions in a function-container (#216)
Related to:
- #200 

Will also fix
- #209 (The ID is now set to be a size_t which is big enough)

To allow for callback deletion, I decided to implement a
FunctionContainer utility. I have used it in all the places where we
store a list of functions, which includes:
- Systems
- InputCallbacks
- ContactCallbacks

Notes:
- Even with the new ResourceManager, it seems that the FunctionContainer
needs to be created as a shared_ptr, otherwise it tries to copy the list
of unique pointers somehow (if there is a better solution let me know)

---------

+++ src/plugin/opengl/src/system/WindowSystems.cpp
@@ -74,8 +74,8 @@ void ES::Plugin::OpenGL::System::SetupMouseDragging(ES::Engine::Core &core)
         else if (ES::Plugin::Input::Utils::IsMouseButtonPressed(GLFW_MOUSE_BUTTON_RIGHT))
             handleRightDrag(camera, dragging, xpos, ypos);
 
-        dragging.lastMousePos.x = xpos;
-        dragging.lastMousePos.y = ypos;
+        dragging.lastMousePos.x = static_cast<float>(xpos);
+        dragging.lastMousePos.y = static_cast<float>(ypos);
     });
 }
